requirements.txt
python-dotenv==0.15.0
python-telegram-bot==13.1
requests==2.31.0